# internal/keeper/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Keeper API",
    "description": "API spec for Tacito Square Keeper component, managing agent configurations, communities, and support entities.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Primary API Version 1"
    }
  ],
...
  "paths": {
...
    "/skills": {
      "post": {
        "summary": "Create a new Skill Collection",
        "operationId": "createSkill",
        "tags": [
          "agent-config/skills"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSkillRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created Skill Collection",
            "headers": {
              "Location": {
                "description": "The URI of the newly created skills",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List all Skill Collections",
        "operationId": "listSkills",
        "tags": [
          "agent-config/skills"
        ],
        "responses": {
          "200": {
            "description": "List of Skill Collections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Skill"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/agents/{agent_id}/skills/{skill_id}": {
      "post": {
        "summary": "Attach a Skill Collection to an Agent",
        "operationId": "attachSkillToAgent",
        "tags": [
          "agent-config/skills"
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "description": "The UUID of the Agent",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "description": "The UUID of the Skill Collection",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully attached",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "attached"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Detach a Skill Collection from an Agent",
        "operationId": "detachSkillFromAgent",
        "tags": [
          "agent-config/skills"
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "description": "The UUID of the Agent",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "description": "The UUID of the Skill Collection",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully detached"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/skill-collections": {
      "post": {
        "summary": "Create a new Skill Collection",
        "operationId": "createSkillCollection",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSkillCollectionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created Skill Collection",
            "headers": {
              "Location": {
                "description": "The URI of the newly created skill collections",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "agent-config/skills"
        ]
      },
      "get": {
        "summary": "List all configured Skill Collections",
        "operationId": "listSkillCollections",
        "responses": {
          "200": {
            "description": "List of Skill Collections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkillCollection"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "agent-config/skills"
        ]
      }
    },
...
    "/skill-collections/{id}/resolve": {
      "get": {
        "summary": "Resolve collection skills",
        "operationId": "resolveCollectionSkills",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The UUID of the Skill Collection",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resolved Skills",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Skill"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "agent-config/skills"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSkillRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "web-search"
          },
          "description": {
            "type": "string",
            "example": "Google search and URL extraction tools"
          }
        }
      },
...
      "Skill": {
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended",
              "inactive"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateSkillCollectionRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "general-skills"
          },
          "description": {
            "type": "string",
            "example": "General assistant capabilities"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "example": [
              "550e8400-e29b-41d4-a716-446655440000"
            ]
          }
        }
      },
...
      "SkillCollection": {
        "type": "object",
        "required": [
          "id",
          "name",
          "skills",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "name is required"
          }
        }
      },
...
    }
  }
}